Refactoring code to enable testing
