+++ hoprd-node-1.yaml
@@ -9,7 +9,8 @@ spec:
   version: riga
   secret:
     secretName: hoprd-ku5wzs
-  enableMonitoring: true
+  monitoring:
+    enabled: true
   # announce: true
   # autoRedeemTickets: true
   # networkQualityThreshold: 0.6
